Struct widestring::WideString

pub struct WideString {
    // some fields omitted
}

An owned, mutable "wide" string for windows FFI that is not nul-aware.

WideString is not aware of nul values. Strings may or may not be nul-terminated, and may contain invalid and ill-formed UTF-16 data. These strings are intended to be used with windows FFI functions that directly use string length, where the strings are known to have proper nul-termination already, or where strings are merely being passed through without modification.

WideCString should be used instead if nul-aware strings are required.

WideString can be converted to and from many other string types, including OsString and String, making proper Unicode windows FFI safe and easy.

Examples

The following example constructs a WideString and shows how to convert a WideString to a regular Rust String.

use widestring::WideString;
let v = vec![84u16, 104u16, 101u16]; // 'T' 'h' 'e'
// Create a wide string from the vector
let wstr = WideString::from_vec(v);
// Convert to a rust string!
let rust_str = wstr.to_string_lossy();
assert_eq!(rust_str, "The");

Methods

impl WideString

fn new() -> WideString

Constructs a new empty WideString.

fn from_vec<T: Into<Vec<u16>>>(raw: T) -> WideString

Constructs a WideString from a vector of possibly invalid or ill-formed UTF-16 data.

No checks are made on the contents of the vector.

Examples

use widestring::WideString;
let v = vec![84u16, 104u16, 101u16]; // 'T' 'h' 'e'
// Create a wide string from the vector
let wstr = WideString::from_vec(v);

fn from_str<S: AsRef<OsStr> + ?Sized>(s: &S) -> WideString

Encodes a WideString copy from an OsStr.

This makes a wide string copy of the OsStr. Since OsStr makes no guaruntees that it is valid data, there is no guaruntee that the resulting WideString will be valid UTF-16.

Examples

use widestring::WideString;
let s = "MyString";
// Create a wide string from the string
let wstr = WideString::from_str(s);

assert_eq!(wstr.to_string().unwrap(), s);

unsafe fn from_ptr(p: *const u16, len: usize) -> WideString

Constructs a WideString from a u16 pointer and a length.

The len argument is the number of u16 elements, not the number of bytes.

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len elements.

Panics

Panics if len is greater than 0 but p is a null pointer.

fn with_capacity(capacity: usize) -> WideString

Creates a WideString with the given capacity.

The string will be able to hold exactly capacity partial code units without reallocating. If capacity is set to 0, the string will not initially allocate.

fn capacity(&self) -> usize

Returns the capacity this WideString can hold without reallocating.

fn clear(&mut self)

Truncate the WideString to zero length.

fn reserve(&mut self, additional: usize)

Reserves the capacity for at least additiona more capacity to be inserted in the given WideString.

More space may be reserved to avoid frequent allocations.

fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additiona more capacity to be inserted in the given WideString. Does nothing if the capcity is already sufficient.

Note that the allocator may give more space than is requested. Therefore capacity can not be relied upon to be precisely minimal. Prefer reserve if future insertions are expected.

fn as_wide_str(&self) -> &WideStr

Converts to a WideStr reference.

fn into_vec(self) -> Vec<u16>

Converts the wide string into a Vec<u16>, consuming the string in the process.

fn push<T: AsRef<WideStr>>(&mut self, s: T)

Extends the wide string with the given &WideStr.

No checks are performed on the strings. It is possible to end up nul values inside the string, and it is up to the caller to determine if that is acceptable.

Examples

use widestring::WideString;
let s = "MyString";
let mut wstr = WideString::from_str(s);
let cloned = wstr.clone();
// Push the clone to the end, repeating the string twice.
wstr.push(cloned);

assert_eq!(wstr.to_string().unwrap(), "MyStringMyString");

fn push_slice<T: AsRef<[u16]>>(&mut self, s: T)

Extends the wide string with the given &[u16] slice.

No checks are performed on the strings. It is possible to end up nul values inside the string, and it is up to the caller to determine if that is acceptable.

Examples

use widestring::WideString;
let s = "MyString";
let mut wstr = WideString::from_str(s);
let cloned = wstr.clone();
// Push the clone to the end, repeating the string twice.
wstr.push_slice(cloned);

assert_eq!(wstr.to_string().unwrap(), "MyStringMyString");

fn push_str<T: AsRef<OsStr>>(&mut self, s: T)

Extends the string with the given &OsStr.

No checks are performed on the strings. It is possible to end up nul values inside the string, and it is up to the caller to determine if that is acceptable.

Examples

use widestring::WideString;
let s = "MyString";
let mut wstr = WideString::from_str(s);
// Push the original to the end, repeating the string twice.
wstr.push_str(s);

assert_eq!(wstr.to_string().unwrap(), "MyStringMyString");