widestring::ucstring

Struct U16CString

pub struct U16CString { /* private fields */ }

An owned, mutable C-style 16-bit wide string for FFI that is nul-aware and nul-terminated.

The string slice of a U16CString is U16CStr.

U16CString strings do not have a defined encoding. While it is sometimes assumed that they contain possibly invalid or ill-formed UTF-16 data, they may be used for any wide encoded string.

Nul termination

U16CString is aware of nul (0) values. Unless unchecked conversions are used, all U16CString strings end with a nul-terminator in the underlying buffer and contain no internal nul values. These strings are intended to be used with FFI functions that require nul-terminated strings.

Because of the nul termination requirement, multiple classes methods for provided for construction a U16CString under various scenarios. By default, methods such as from_ptr and from_vec return an error if it contains any interior nul values before the terminator. For these methods, the input does not need to contain the terminating nul; it is added if it is does not exist.

_truncate methods on the other hand, such as from_ptr_truncate and from_vec_truncate, construct a string that terminates with the first nul value encountered in the string, and do not return an error. They automatically ensure the string is terminated in a nul value even if it was not originally.

Finally, unsafe _unchecked variants of these methods, such as from_ptr_unchecked and from_vec_unchecked allow bypassing any checks for nul values, when the input has already been ensured to no interior nul values. Again, any missing nul terminator is automatically added if necessary.

Examples

The easiest way to use U16CString outside of FFI is with the u16cstr! macro to convert string literals into nul-terminated UTF-16 strings at compile time:

use widestring::{u16cstr, U16CString};
let hello = U16CString::from(u16cstr!("Hello, world!"));

You can also convert any u16 slice or vector directly:

use widestring::{u16cstr, U16CString};

let sparkle_heart = vec![0xd83d, 0xdc96];
let sparkle_heart = U16CString::from_vec(sparkle_heart).unwrap();
// The string will add the missing nul terminator

assert_eq!(u16cstr!("๐"), sparkle_heart);

// This unpaired UTf-16 surrogate is invalid UTF-16, but is perfectly valid in U16CString
let malformed_utf16 = vec![0xd83d, 0x0];
let s = U16CString::from_vec(malformed_utf16).unwrap();

assert_eq!(s.len(), 1); // Note the terminating nul is not counted in the length

When working with a FFI, it is useful to create a U16CString from a pointer:

use widestring::{u16cstr, U16CString};

let sparkle_heart = [0xd83d, 0xdc96, 0x0];
let s = unsafe {
    // Note the string and pointer length does not include the nul terminator
    U16CString::from_ptr(sparkle_heart.as_ptr(), sparkle_heart.len() - 1).unwrap()
};
assert_eq!(u16cstr!("๐"), s);

// Alternatively, if the length of the pointer is unknown but definitely terminates in nul,
// a C-style string version can be used
let s = unsafe { U16CString::from_ptr_str(sparkle_heart.as_ptr()) };

assert_eq!(u16cstr!("๐"), s);

Implementations

impl U16CString

pub const NUL_TERMINATOR: u16 = 0u16

The nul terminator character value.

pub fn new() -> Self

Constructs a new empty wide C string.

pub fn from_vec(v: impl Into<Vec<u16>>) -> Result<Self, ContainsNul<u16>>

Constructs a wide C string from a container of wide character data.

This method will consume the provided data and use the underlying elements to construct a new string. The data will be scanned for invalid interior nul values.

Errors

This function will return an error if the data contains a nul value that is not the terminating nul. The returned error will contain the original Vec as well as the position of the nul value.

Examples

use widestring::U16CString;
let v = vec![84u16, 104u16, 101u16]; // 'T' 'h' 'e'
// Create a wide string from the vector
let wcstr = U16CString::from_vec(v).unwrap();

Empty vectors are valid and will return an empty string with a nul terminator:

use widestring::U16CString;
let wcstr = U16CString::from_vec(vec![]).unwrap();
assert_eq!(wcstr, U16CString::default());

The following example demonstrates errors from nul values in a vector.

use widestring::U16CString;
let v = vec![84u16, 0u16, 104u16, 101u16]; // 'T' NUL 'h' 'e'
// Create a wide string from the vector
let res = U16CString::from_vec(v);
assert!(res.is_err());
assert_eq!(res.err().unwrap().nul_position(), 1);

pub fn from_vec_truncate(v: impl Into<Vec<u16>>) -> Self

Constructs a wide C string from a container of wide character data, truncating at the first nul terminator.

The string will be truncated at the first nul value in the data.

Examples

use widestring::U16CString;
let v = vec![84u16, 104u16, 101u16, 0u16]; // 'T' 'h' 'e' NUL
// Create a wide string from the vector
let wcstr = U16CString::from_vec_truncate(v);

pub unsafe fn from_vec_unchecked(v: impl Into<Vec<u16>>) -> Self

Constructs a wide C string from a vector without checking for interior nul values.

A terminating nul value will be appended if the vector does not already have a terminating nul.

Safety

This method is equivalent to from_vec except that no runtime assertion is made that v contains no interior nul values. Providing a vector with any nul values that are not the last value in the vector will result in an invalid C string.

pub fn from_ustr(s: impl AsRef<U16Str>) -> Result<Self, ContainsNul<u16>>

Constructs a wide C string from anything that can be converted to a wide string slice.

The string will be scanned for invalid interior nul values.

Errors

This function will return an error if the data contains a nul value that is not the terminating nul. The returned error will contain a Vec as well as the position of the nul value.

pub fn from_ustr_truncate(s: impl AsRef<U16Str>) -> Self

Constructs a wide C string from anything that can be converted to a wide string slice, truncating at the first nul terminator.

The string will be truncated at the first nul value in the string.

pub unsafe fn from_ustr_unchecked(s: impl AsRef<U16Str>) -> Self

Constructs a wide C string from anything that can be converted to a wide string slice, without scanning for invalid nul values.

Safety

This method is equivalent to from_ustr except that no runtime assertion is made that v contains no interior nul values. Providing a string with any nul values that are not the last value in the vector will result in an invalid C string.

pub unsafe fn from_ptr_str(p: *const u16) -> Self

Constructs a new wide C string copied from a nul-terminated string pointer.

This will scan for nul values beginning with p. The first nul value will be used as the nul terminator for the string, similar to how libc string functions such as strlen work.

If you wish to avoid copying the string pointer, use U16CStr::from_ptr_str or U32CStr::from_ptr_str instead.

Safety

This function is unsafe as there is no guarantee that the given pointer is valid or has a nul terminator, and the function could scan past the underlying buffer.

In addition, the data must meet the safety conditions of std::slice::from_raw_parts.

Panics

This function panics if p is null.

Caveat

The lifetime for the returned string is inferred from its usage. To prevent accidental misuse, itโs suggested to tie the lifetime to whichever source lifetime is safe in the context, such as by providing a helper function taking the lifetime of a host value for the string, or by explicit annotation.

pub unsafe fn from_ptr( p: *const u16, len: usize, ) -> Result<Self, ContainsNul<u16>>

Constructs a wide C string copied from a pointer and a length, checking for invalid interior nul values.

The len argument is the number of elements, not the number of bytes, and does not include the nul terminator of the string. If len is 0, p is allowed to be a null pointer.

The resulting string will always be nul-terminated even if the pointer data is not.

Errors

This will scan the pointer string for an interior nul value and error if one is found. To avoid scanning for interior nuls, from_ptr_unchecked may be used instead. The returned error will contain a Vec as well as the position of the nul value.

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len elements.

In addition, the data must meet the safety conditions of std::slice::from_raw_parts.

Panics

Panics if len is greater than 0 but p is a null pointer.

pub unsafe fn from_ptr_truncate(p: *const u16, len: usize) -> Self

Constructs a wide C string copied from a pointer and a length, truncating at the first nul terminator.

The len argument is the number of elements, not the number of bytes. This will scan for nul values beginning with p until offset len. The first nul value will be used as the nul terminator for the string, ignoring any remaining values left before len. If no nul value is found, the whole string of length len is used, and a new nul-terminator will be added to the resulting string. If len is 0, p is allowed to be a null pointer.

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len elements.

In addition, the data must meet the safety conditions of std::slice::from_raw_parts.

Panics

Panics if len is greater than 0 but p is a null pointer.

pub unsafe fn from_ptr_unchecked(p: *const u16, len: usize) -> Self

Constructs a wide C string copied from a pointer and a length without checking for any nul values.

The len argument is the number of elements, not the number of bytes, and does not include the nul terminator of the string. If len is 0, p is allowed to be a null pointer.

The resulting string will always be nul-terminated even if the pointer data is not.

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len elements.

In addition, the data must meet the safety conditions of std::slice::from_raw_parts.

The interior values of the pointer are not scanned for nul. Any interior nul values or will result in an invalid C string.

Panics

Panics if len is greater than 0 but p is a null pointer.

pub fn as_ucstr(&self) -> &U16CStr

Converts to a wide C string slice.

pub fn as_mut_ucstr(&mut self) -> &mut U16CStr

Converts to a mutable wide C string slice.

pub fn into_ustring(self) -> U16String

Converts this string into a wide string without a nul terminator.

The resulting string will not contain a nul-terminator, and will contain no other nul values.

pub fn into_ustring_with_nul(self) -> U16String

Converts this string into a wide string with a nul terminator.

The resulting vector will contain a nul-terminator and no interior nul values.

pub fn into_vec(self) -> Vec<u16>

Converts the string into a Vec without a nul terminator, consuming the string in the process.

The resulting vector will not contain a nul-terminator, and will contain no other nul values.

pub fn into_vec_with_nul(self) -> Vec<u16>

Converts the string into a Vec, consuming the string in the process.

The resulting vector will contain a nul-terminator and no interior nul values.

pub fn into_raw(self) -> *mut u16

Transfers ownership of the string to a C caller.

Safety

The pointer must be returned to Rust and reconstituted using from_raw to be properly deallocated. Specifically, one should not use the standard C free function to deallocate this string. Failure to call from_raw will lead to a memory leak.

pub unsafe fn from_raw(p: *mut u16) -> Self

Retakes ownership of a wide C string that was transferred to C.

This should only be used in combination with into_raw. To construct a new wide C string from a pointer, use from_ptr_str.

Safety

This should only ever be called with a pointer that was earlier obtained by calling into_raw. Additionally, the length of the string will be recalculated from the pointer by scanning for the nul-terminator.

Panics

Panics if p is a null pointer.

pub fn into_boxed_ucstr(self) -> Box<U16CStr>

Converts this wide C string into a boxed wide C string slice.

Examples

use widestring::{U16CString, U16CStr};

let mut v = vec![102u16, 111u16, 111u16]; // "foo"
let c_string = U16CString::from_vec(v.clone()).unwrap();
let boxed = c_string.into_boxed_ucstr();
v.push(0);
assert_eq!(&*boxed, U16CStr::from_slice(&v).unwrap());

impl U16CString

pub fn from_str(s: impl AsRef<str>) -> Result<Self, ContainsNul<u16>>

Constructs a U16CString copy from a str, encoding it as UTF-16.

This makes a string copy of the str. Since str will always be valid UTF-8, the resulting U16CString will also be valid UTF-16.

The string will be scanned for nul values, which are invalid anywhere except the final character.

The resulting string will always be nul-terminated even if the original string is not.

Errors

This function will return an error if the data contains a nul value anywhere except the final position. The returned error will contain a Vec<u16> as well as the position of the nul value.

Examples

use widestring::U16CString;
let s = "MyString";
// Create a wide string from the string
let wcstr = U16CString::from_str(s).unwrap();

The following example demonstrates errors from nul values in a string.

use widestring::U16CString;
let s = "My\u{0}String";
// Create a wide string from the string
let res = U16CString::from_str(s);
assert!(res.is_err());
assert_eq!(res.err().unwrap().nul_position(), 2);

pub unsafe fn from_str_unchecked(s: impl AsRef<str>) -> Self

Constructs a U16CString copy from a str, encoding it as UTF-16, without checking for interior nul values.

This makes a string copy of the str. Since str will always be valid UTF-8, the resulting U16CString will also be valid UTF-16.

The resulting string will always be nul-terminated even if the original string is not.

Safety

This method is equivalent to from_str except that no runtime assertion is made that s contains no interior nul values. Providing a string with nul values that are not the last character will result in an invalid U16CString.

Examples

use widestring::U16CString;
let s = "MyString";
// Create a wide string from the string
let wcstr = unsafe { U16CString::from_str_unchecked(s) };

pub fn from_str_truncate(s: impl AsRef<str>) -> Self

Constructs a U16CString copy from a str, encoding it as UTF-16, truncating at the first nul terminator.

This makes a string copy of the str. Since str will always be valid UTF-8, the resulting U16CString will also be valid UTF-16.

The string will be truncated at the first nul value in the string. The resulting string will always be nul-terminated even if the original string is not.

Examples

use widestring::U16CString;
let s = "My\u{0}String";
// Create a wide string from the string
let wcstr = U16CString::from_str_truncate(s);
assert_eq!(wcstr.to_string_lossy(), "My");

pub fn from_os_str(s: impl AsRef<OsStr>) -> Result<Self, ContainsNul<u16>>

Constructs a U16CString copy from an OsStr.

This makes a string copy of the OsStr. Since OsStr makes no guarantees that it is valid data, there is no guarantee that the resulting U16CString will be valid UTF-16.

The string will be scanned for nul values, which are invalid anywhere except the final character. The resulting string will always be nul-terminated even if the original string is not.

Note that the encoding of OsStr is platform-dependent, so on some platforms this may make an encoding conversions, while on other platforms (such as windows) no changes to the string will be made.

Errors

This function will return an error if the data contains a nul value anywhere except the last character. The returned error will contain a Vec<u16> as well as the position of the nul value.

Examples

use widestring::U16CString;
let s = "MyString";
// Create a wide string from the string
let wcstr = U16CString::from_os_str(s).unwrap();

The following example demonstrates errors from nul values in the string.

use widestring::U16CString;
let s = "My\u{0}String";
// Create a wide string from the string
let res = U16CString::from_os_str(s);
assert!(res.is_err());
assert_eq!(res.err().unwrap().nul_position(), 2);

pub unsafe fn from_os_str_unchecked(s: impl AsRef<OsStr>) -> Self

Constructs a U16CString copy from an OsStr, without checking for nul values.

This makes a string copy of the OsStr. Since OsStr makes no guarantees that it is valid data, there is no guarantee that the resulting U16CString will be valid UTF-16.

The resulting string will always be nul-terminated even if the original string is not.

Note that the encoding of OsStr is platform-dependent, so on some platforms this may make an encoding conversions, while on other platforms (such as windows) no changes to the string will be made.

Safety

This method is equivalent to from_os_str except that no runtime assertion is made that s contains no interior nul values. Providing a string with nul values anywhere but the last character will result in an invalid U16CString.

Examples

use widestring::U16CString;
let s = "MyString";
// Create a wide string from the string
let wcstr = unsafe { U16CString::from_os_str_unchecked(s) };

pub fn from_os_str_truncate(s: impl AsRef<OsStr>) -> Self

Constructs a U16CString copy from an OsStr, truncating at the first nul terminator.

This makes a string copy of the OsStr. Since OsStr makes no guarantees that it is valid data, there is no guarantee that the resulting U16CString will be valid UTF-16.

The string will be truncated at the first nul value in the string. The resulting string will always be nul-terminated even if the original string is not.

Note that the encoding of OsStr is platform-dependent, so on some platforms this may make an encoding conversions, while on other platforms (such as windows) no changes to the string will be made.

Examples

use widestring::U16CString;
let s = "My\u{0}String";
// Create a wide string from the string
let wcstr = U16CString::from_os_str_truncate(s);
assert_eq!(wcstr.to_string_lossy(), "My");