Struct widestring::ucstring::UCString

pub struct UCString<C: UChar> { /* fields omitted */ }

An owned, mutable C-style “wide” string for FFI that is nul-aware and nul-terminated.

UCString is aware of nul values. Unless unchecked conversions are used, all UCString strings end with a nul-terminator in the underlying buffer and contain no internal nul values. The strings may still contain invalid or ill-formed UTF-16 or UTF-32 data. These strings are intended to be used with FFI functions such as Windows API that may require nul-terminated strings.

UCString can be converted to and from many other string types, including UString, OsString, and String, making proper Unicode FFI safe and easy.

Please prefer using the type aliases U16CString, U32CString, or WideCString to using this type directly.

Examples

The following example constructs a U16CString and shows how to convert a U16CString to a regular Rust String.

use widestring::U16CString;
let s = "Test";
// Create a wide string from the rust string
let wstr = U16CString::from_str(s).unwrap();
// Convert back to a rust string
let rust_str = wstr.to_string_lossy();
assert_eq!(rust_str, "Test");

The same example using U32CString:

use widestring::U32CString;
let s = "Test";
// Create a wide string from the rust string
let wstr = U32CString::from_str(s).unwrap();
// Convert back to a rust string
let rust_str = wstr.to_string_lossy();
assert_eq!(rust_str, "Test");

Implementations

impl<C: UChar> UCString<C>

pub fn from_vec(v: impl Into<Vec<C>>) -> Result<Self, ContainsNul<C>>

Constructs a UCString from a container of wide character data.

This method will consume the provided data and use the underlying elements to construct a new string. The data will be scanned for invalid interior nul values.

Errors

This function will return an error if the data contains a nul value that is not the terminating nul. The returned error will contain the original Vec as well as the position of the nul value.

Examples

use widestring::U16CString;
let v = vec![84u16, 104u16, 101u16]; // 'T' 'h' 'e'
// Create a wide string from the vector
let wcstr = U16CString::from_vec(v).unwrap();

use widestring::U32CString;
let v = vec![84u32, 104u32, 101u32]; // 'T' 'h' 'e'
// Create a wide string from the vector
let wcstr = U32CString::from_vec(v).unwrap();

Empty vectors are valid and will return an empty string with a nul terminator:

use widestring::U16CString;
let wcstr = U16CString::from_vec(vec![]).unwrap();
assert_eq!(wcstr, U16CString::default());

use widestring::U32CString;
let wcstr = U32CString::from_vec(vec![]).unwrap();
assert_eq!(wcstr, U32CString::default());

The following example demonstrates errors from nul values in a vector.

use widestring::U16CString;
let v = vec![84u16, 0u16, 104u16, 101u16]; // 'T' NUL 'h' 'e'
// Create a wide string from the vector
let res = U16CString::from_vec(v);
assert!(res.is_err());
assert_eq!(res.err().unwrap().nul_position(), 1);

use widestring::U32CString;
let v = vec![84u32, 0u32, 104u32, 101u32]; // 'T' NUL 'h' 'e'
// Create a wide string from the vector
let res = U32CString::from_vec(v);
assert!(res.is_err());
assert_eq!(res.err().unwrap().nul_position(), 1);

pub fn from_vec_truncate(v: impl Into<Vec<C>>) -> Self

Constructs a UCString from a container of wide character data, truncating at the first nul terminator.

The string will be truncated at the first nul value in the data.

Examples

use widestring::U16CString;
let v = vec![84u16, 104u16, 101u16, 0u16]; // 'T' 'h' 'e' NUL
// Create a wide string from the vector
let wcstr = U16CString::from_vec_truncate(v);

use widestring::U32CString;
let v = vec![84u32, 104u32, 101u32, 0u32]; // 'T' 'h' 'e' NUL
// Create a wide string from the vector
let wcstr = U32CString::from_vec_truncate(v);

pub unsafe fn from_vec_unchecked(v: impl Into<Vec<C>>) -> Self

Constructs a UCString from a vector without checking for interior nul values.

A terminating nul value will be appended if the vector does not already have a terminating nul.

Safety

This method is equivalent to from_vec except that no runtime assertion is made that v contains no interior nul values. Providing a vector with any nul values that are not the last value in the vector will result in an invalid UCString.

pub fn from_ustr(s: impl AsRef<UStr<C>>) -> Result<Self, ContainsNul<C>>

Constructs a UCString from anything that can be converted to a UStr.

The string will be scanned for invalid interior nul values.

Errors

This function will return an error if the data contains a nul value that is not the terminating nul. The returned error will contain a Vec as well as the position of the nul value.

pub fn from_ustr_truncate(s: impl AsRef<UStr<C>>) -> Self

Constructs a UCString from anything that can be converted to a UStr, truncating at the first nul terminator.

The string will be truncated at the first nul value in the string.

pub unsafe fn from_ustr_unchecked(s: impl AsRef<UStr<C>>) -> Self

Constructs a UCString from anything that can be converted to a UStr, without scanning for invalid nul values.

Safety

This method is equivalent to from_ustr except that no runtime assertion is made that v contains no interior nul values. Providing a string with any nul values that are not the last value in the vector will result in an invalid UCString.

pub unsafe fn from_ptr_str(p: *const C) -> Self

Constructs a new UCString copied from a nul-terminated string pointer.

This will scan for nul values beginning with p. The first nul value will be used as the nul terminator for the string, similar to how libc string functions such as strlen work.

If you wish to avoid copying the string pointer, use UCStr::from_ptr_str instead.

Safety

This function is unsafe as there is no guarantee that the given pointer is valid or has a nul terminator, and the function could scan past the underlying buffer.

In addition, the data must meet the safety conditions of std::slice::from_raw_parts.

Panics

This function panics if p is null.

Caveat

The lifetime for the returned string is inferred from its usage. To prevent accidental misuse, it’s suggested to tie the lifetime to whichever source lifetime is safe in the context, such as by providing a helper function taking the lifetime of a host value for the string, or by explicit annotation.

pub unsafe fn from_ptr(p: *const C, len: usize) -> Result<Self, ContainsNul<C>>

Constructs a UCString copied from a pointer and a length, checking for invalid interior nul values.

The len argument is the number of elements, not the number of bytes, and does not include the nul terminator of the string. If len is 0, p is allowed to be a null pointer.

The resulting string will always be nul-terminated even if the pointer data is not.

Errors

This will scan the pointer string for an interior nul value and error if one is found. To avoid scanning for interior nuls, from_ptr_unchecked may be used instead. The returned error will contain a Vec as well as the position of the nul value.

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len elements.

In addition, the data must meet the safety conditions of std::slice::from_raw_parts.

Panics

Panics if len is greater than 0 but p is a null pointer.

pub unsafe fn from_ptr_truncate(p: *const C, len: usize) -> Self

Constructs a UCString copied from a pointer and a length, truncating at the first nul terminator.

The len argument is the number of elements, not the number of bytes. This will scan for nul values beginning with p until offset len. The first nul value will be used as the nul terminator for the string, ignoring any remaining values left before len. If no nul value is found, the whole string of length len is used, and a new nul-terminator will be added to the resulting string. If len is 0, p is allowed to be a null pointer.

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len elements.

In addition, the data must meet the safety conditions of std::slice::from_raw_parts.

Panics

Panics if len is greater than 0 but p is a null pointer.

pub unsafe fn from_ptr_unchecked(p: *const C, len: usize) -> Self

Constructs a UCString copied from a pointer and a length without checking for any nul values.

The len argument is the number of elements, not the number of bytes, and does not include the nul terminator of the string. If len is 0, p is allowed to be a null pointer.

The resulting string will always be nul-terminated even if the pointer data is not.

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len elements.

In addition, the data must meet the safety conditions of std::slice::from_raw_parts.

The interior values of the pointer are not scanned for nul. Any interior nul values or will result in an invalid UCString.

Panics

Panics if len is greater than 0 but p is a null pointer.

pub fn as_ucstr(&self) -> &UCStr<C>

Converts to a UCStr reference.

pub fn as_mut_ucstr(&mut self) -> &mut UCStr<C>

Converts to a mutable UCStr reference.

pub fn into_ustring(self) -> UString<C>

Converts this string into a UString without a nul terminator.

The resulting string will not contain a nul-terminator, and will contain no other nul values.

pub fn into_ustring_with_nul(self) -> UString<C>

Converts this string into a UString with a nul terminator.

The resulting vector will contain a nul-terminator and no interior nul values.

pub fn into_vec(self) -> Vec<C>

Converts the string into a Vec without a nul terminator, consuming the string in the process.

The resulting vector will not contain a nul-terminator, and will contain no other nul values.

pub fn into_vec_with_nul(self) -> Vec<C>

Converts the string into a Vec, consuming the string in the process.

The resulting vector will contain a nul-terminator and no interior nul values.

pub fn into_raw(self) -> *mut C

Transfers ownership of the string to a C caller.

Safety

The pointer must be returned to Rust and reconstituted using from_raw to be properly deallocated. Specifically, one should not use the standard C free function to deallocate this string. Failure to call from_raw will lead to a memory leak.

pub unsafe fn from_raw(p: *mut C) -> Self

Retakes ownership of a UCString that was transferred to C.

This should only be used in combination with into_raw. To construct a new UCString from a pointer, use from_ptr_str.

Safety

This should only ever be called with a pointer that was earlier obtained by calling into_raw. Additionally, the length of the string will be recalculated from the pointer by scanning for the nul-terminator.

Panics

Panics if p is a null pointer.

pub fn into_boxed_ucstr(self) -> Box<UCStr<C>>

Converts this UCString into a boxed UCStr.

Examples

use widestring::{U16CString, U16CStr};

let mut v = vec![102u16, 111u16, 111u16]; // "foo"
let c_string = U16CString::from_vec(v.clone()).unwrap();
let boxed = c_string.into_boxed_ucstr();
v.push(0);
assert_eq!(&*boxed, U16CStr::from_slice(&v).unwrap());

use widestring::{U32CString, U32CStr};

let mut v = vec![102u32, 111u32, 111u32]; // "foo"
let c_string = U32CString::from_vec(v.clone()).unwrap();
let boxed = c_string.into_boxed_ucstr();
v.push(0);
assert_eq!(&*boxed, U32CStr::from_slice(&v).unwrap());

impl UCString<u16>

pub fn from_str(s: impl AsRef<str>) -> Result<Self, ContainsNul<u16>>

Encodes a U16CString copied from a str.

The string will be scanned for nul values, which are invalid anywhere except the final character.

The resulting string will always be nul-terminated even if the original string is not.

Errors

This function will return an error if the data contains a nul value anywhere except the final position. The returned error will contain a Vec<u16> as well as the position of the nul value.

Examples

use widestring::U16CString;
let s = "MyString";
// Create a wide string from the string
let wcstr = U16CString::from_str(s).unwrap();

The following example demonstrates errors from nul values in a string.

use widestring::U16CString;
let s = "My\u{0}String";
// Create a wide string from the string
let res = U16CString::from_str(s);
assert!(res.is_err());
assert_eq!(res.err().unwrap().nul_position(), 2);

pub unsafe fn from_str_unchecked(s: impl AsRef<str>) -> Self

Encodes a U16CString copied from a str, without checking for interior nul values.

The resulting string will always be nul-terminated even if the original string is not.

Safety

This method is equivalent to from_str except that no runtime assertion is made that s contains no interior nul values. Providing a string with nul values that are not the last character will result in an invalid U16CString.

Examples

use widestring::U16CString;
let s = "MyString";
// Create a wide string from the string
let wcstr = unsafe { U16CString::from_str_unchecked(s) };

pub fn from_str_truncate(s: impl AsRef<str>) -> Self

Encodes a U16CString copied from a str, truncating at the first nul terminator.

The string will be truncated at the first nul value in the string. The resulting string will always be nul-terminated even if the original string is not.

Examples

use widestring::U16CString;
let s = "My\u{0}String";
// Create a wide string from the string
let wcstr = U16CString::from_str_truncate(s);
assert_eq!(wcstr.to_string_lossy(), "My");

pub fn from_os_str(s: impl AsRef<OsStr>) -> Result<Self, ContainsNul<u16>>

Encodes a U16CString copied from anything that can be converted to an OsStr.

The string will be scanned for nul values, which are invalid anywhere except the final character. The resulting string will always be nul-terminated even if the original string is not.

Note that the encoding of OsStr is platform-dependent, so on some platforms this may make an encoding conversions, while on other platforms (such as windows) no changes to the string will be made.

Errors

This function will return an error if the data contains a nul value anywhere except the last character. The returned error will contain a Vec<u16> as well as the position of the nul value.

Examples

use widestring::U16CString;
let s = "MyString";
// Create a wide string from the string
let wcstr = U16CString::from_os_str(s).unwrap();

The following example demonstrates errors from nul values in the string.

use widestring::U16CString;
let s = "My\u{0}String";
// Create a wide string from the string
let res = U16CString::from_os_str(s);
assert!(res.is_err());
assert_eq!(res.err().unwrap().nul_position(), 2);

pub unsafe fn from_os_str_unchecked(s: impl AsRef<OsStr>) -> Self

Encodes a U16CString from anything that can be converted to an OsStr, without checking for nul values.

The resulting string will always be nul-terminated even if the original string is not.

Note that the encoding of OsStr is platform-dependent, so on some platforms this may make an encoding conversions, while on other platforms (such as windows) no changes to the string will be made.

Safety

This method is equivalent to from_os_str except that no runtime assertion is made that s contains no interior nul values. Providing a string with nul values anywhere but the last character will result in an invalid U16CString.

Examples

use widestring::U16CString;
let s = "MyString";
// Create a wide string from the string
let wcstr = unsafe { U16CString::from_os_str_unchecked(s) };

pub fn from_os_str_truncate(s: impl AsRef<OsStr>) -> Self

Encodes a U16CString copied from anything that can be converted to an OsStr, truncating at the first nul terminator.

The string will be truncated at the first nul value in the string. The resulting string will always be nul-terminated even if the original string is not.

Note that the encoding of OsStr is platform-dependent, so on some platforms this may make an encoding conversions, while on other platforms (such as windows) no changes to the string will be made.

Examples

use widestring::U16CString;
let s = "My\u{0}String";
// Create a wide string from the string
let wcstr = U16CString::from_os_str_truncate(s);
assert_eq!(wcstr.to_string_lossy(), "My");

impl UCString<u32>

pub fn from_chars(v: impl Into<Vec<char>>) -> Result<Self, ContainsNul<u32>>

Constructs a U32CString from a container of character data, checking for invalid nul values.

This method will consume the provided data and use the underlying elements to construct a new string. The data will be scanned for invalid nul values anywhere except the last character. The resulting string will always be nul-terminated even if the original string is not.

Errors

This function will return an error if the data contains a nul value anywhere except the last character. The returned error will contain the Vec<u32> as well as the position of the nul value.

Examples

use widestring::U32CString;
let v: Vec<char> = "Test".chars().collect();
// Create a wide string from the vector
let wcstr = U32CString::from_chars(v).unwrap();

The following example demonstrates errors from nul values in a vector.

use widestring::U32CString;
let v: Vec<char> = "T\u{0}est".chars().collect();
// Create a wide string from the vector
let res = U32CString::from_chars(v);
assert!(res.is_err());
assert_eq!(res.err().unwrap().nul_position(), 1);

pub fn from_chars_truncate(v: impl Into<Vec<char>>) -> Self

Constructs a U32CString from a container of character data, truncating at the first nul value.

This method will consume the provided data and use the underlying elements to construct a new string. The string will be truncated at the first nul value in the string. The resulting string will always be nul-terminated even if the original string is not.

Examples

use widestring::U32CString;
let v: Vec<char> = "Test\u{0}".chars().collect();
// Create a wide string from the vector
let wcstr = U32CString::from_chars_truncate(v);

pub unsafe fn from_chars_unchecked(v: impl Into<Vec<char>>) -> Self

Constructs a U32CString from character data without checking for nul values.

A terminating nul value will be appended if the vector does not already have a terminating nul.

Safety

This method is equivalent to from_chars except that no runtime assertion is made that v contains no interior nul values. Providing a vector with nul values anywhere but the last character will result in an invalid U32CString.

pub fn from_str(s: impl AsRef<str>) -> Result<Self, ContainsNul<u32>>

Encodes a U32CString copied from a str, checking for invalid interior nul values.

The string will be scanned for nul values, which are invalid anywhere except the last character. The resulting string will always be nul-terminated even if the original string is not.

Errors

This function will return an error if the data contains a nul value anywhere except the last character. The returned error will contain a Vec<u32> as well as the position of the nul value.

Examples

use widestring::U32CString;
let s = "MyString";
// Create a wide string from the string
let wcstr = U32CString::from_str(s).unwrap();

The following example demonstrates errors from nul values in a string.

use widestring::U32CString;
let s = "My\u{0}String";
// Create a wide string from the string
let res = U32CString::from_str(s);
assert!(res.is_err());
assert_eq!(res.err().unwrap().nul_position(), 2);

pub unsafe fn from_str_unchecked(s: impl AsRef<str>) -> Self

Encodes a U32CString copied from a str, without checking for nul values.

The resulting string will always be nul-terminated even if the original string is not.

Safety

This method is equivalent to from_str except that no runtime assertion is made that s contains invalid nul values. Providing a string with nul values anywhere except the last character will result in an invalid U32CString.

Examples

use widestring::U32CString;
let s = "MyString";
// Create a wide string from the string
let wcstr = unsafe { U32CString::from_str_unchecked(s) };

pub fn from_str_truncate(s: impl AsRef<str>) -> Self

Encodes a U32CString copied from a str, truncating at the first nul terminator.

The string will be truncated at the first nul value in the string. The resulting string will always be nul-terminated even if the original string is not.

Examples

use widestring::U32CString;
let s = "My\u{0}String";
// Create a wide string from the string
let wcstr = U32CString::from_str_truncate(s);
assert_eq!(wcstr.to_string_lossy(), "My");

pub unsafe fn from_char_ptr_str(p: *const char) -> Self

Constructs a new UCString copied from a nul-terminated char string pointer.

This will scan for nul values beginning with p. The first nul value will be used as the nul terminator for the string, similar to how libc string functions such as strlen work.

If you wish to avoid copying the string pointer, use UCStr::from_char_ptr_str instead.

Safety

This function is unsafe as there is no guarantee that the given pointer is valid or has a nul terminator, and the function could scan past the underlying buffer.

In addition, the data must meet the safety conditions of std::slice::from_raw_parts.

Panics

This function panics if p is null.

Caveat

The lifetime for the returned string is inferred from its usage. To prevent accidental misuse, it’s suggested to tie the lifetime to whichever source lifetime is safe in the context, such as by providing a helper function taking the lifetime of a host value for the string, or by explicit annotation.

pub unsafe fn from_char_ptr(
p: *const char,
len: usize
) -> Result<Self, ContainsNul<u32>>

Constructs a UCString copied from a char pointer and a length, checking for invalid interior nul values.

The len argument is the number of elements, not the number of bytes, and does not include the nul terminator of the string. If len is 0, p is allowed to be a null pointer.

The resulting string will always be nul-terminated even if the pointer data is not.

Errors

This will scan the pointer string for an interior nul value and error if one is found. To avoid scanning for interior nuls, from_ptr_unchecked may be used instead. The returned error will contain a Vec as well as the position of the nul value.

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len elements.

In addition, the data must meet the safety conditions of std::slice::from_raw_parts.

Panics

Panics if len is greater than 0 but p is a null pointer.

pub unsafe fn from_char_ptr_truncate(p: *const char, len: usize) -> Self

Constructs a UCString copied from a char pointer and a length, truncating at the first nul terminator.

The len argument is the number of elements, not the number of bytes. This will scan for nul values beginning with p until offset len. The first nul value will be used as the nul terminator for the string, ignoring any remaining values left before len. If no nul value is found, the whole string of length len is used, and a new nul-terminator will be added to the resulting string. If len is 0, p is allowed to be a null pointer.

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len elements.

In addition, the data must meet the safety conditions of std::slice::from_raw_parts.

Panics

Panics if len is greater than 0 but p is a null pointer.

pub unsafe fn from_char_ptr_unchecked(p: *const char, len: usize) -> Self

Constructs a UCString copied from a char pointer and a length without checking for any nul values.

The len argument is the number of elements, not the number of bytes, and does not include the nul terminator of the string. If len is 0, p is allowed to be a null pointer.

The resulting string will always be nul-terminated even if the pointer data is not.

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len elements.

In addition, the data must meet the safety conditions of std::slice::from_raw_parts.

The interior values of the pointer are not scanned for nul. Any interior nul values or will result in an invalid UCString.

Panics

Panics if len is greater than 0 but p is a null pointer.

pub fn from_os_str(s: impl AsRef<OsStr>) -> Result<Self, ContainsNul<u32>>

Encodes a U32CString copied from anything that can be converted to an OsStr, checking for invalid nul values.

The string will be scanned for nul values, which are invlaid anywhere except the last character. The resulting string will always be nul-terminated even if the string is not.

Note that the encoding of OsStr is platform-dependent, so on some platforms this may make an encoding conversions, while on other platforms no changes to the string will be made.

Errors

This function will return an error if the data contains a nul value anywhere except the last character. The returned error will contain a Vec<u16> as well as the position of the nul value.

Examples

use widestring::U32CString;
let s = "MyString";
// Create a wide string from the string
let wcstr = U32CString::from_os_str(s).unwrap();

The following example demonstrates errors from nul values in a string.

use widestring::U32CString;
let s = "My\u{0}String";
// Create a wide string from the string
let res = U32CString::from_os_str(s);
assert!(res.is_err());
assert_eq!(res.err().unwrap().nul_position(), 2);

pub unsafe fn from_os_str_unchecked(s: impl AsRef<OsStr>) -> Self

Encodes a U32CString copied from anything that can be converted to an OsStr, without checking for nul values.

The resulting string will always be nul-terminated even if the string is not.

Note that the encoding of OsStr is platform-dependent, so on some platforms this may make an encoding conversions, while on other platforms no changes to the string will be made.

Safety

This method is equivalent to from_os_str except that no runtime assertion is made that s contains invalid nul values. Providing a string with nul values anywhere except the last character will result in an invalid U32CString.

Examples

use widestring::U32CString;
let s = "MyString";
// Create a wide string from the string
let wcstr = unsafe { U32CString::from_os_str_unchecked(s) };

pub fn from_os_str_truncate(s: impl AsRef<OsStr>) -> Self

Encodes a U32CString copied from anything that can be converted to an OsStr, truncating at the first nul terminator.

The string will be truncated at the first nul value in the string. The resulting string will always be nul-terminated even if the string is not.

Note that the encoding of OsStr is platform-dependent, so on some platforms this may make an encoding conversions, while on other platforms no changes to the string will be made.

Examples

use widestring::U32CString;
let s = "My\u{0}String";
// Create a wide string from the string
let wcstr = U32CString::from_os_str_truncate(s);
assert_eq!(wcstr.to_string_lossy(), "My");