Module widestring::ustr

Wide string slices with undefined encoding.

This module contains wide string slices and related types.

Structs

CharIndicesLossyUtf16

A lossy iterator over the chars of a string slice, and their positions.

CharIndicesLossyUtf32

A lossy iterator over the chars of a string slice, and their positions.

CharIndicesUtf16

An iterator over the decoded chars of a string slice, and their positions.

CharIndicesUtf32

An iterator over the decoded chars of a string slice, and their positions.

CharsLossyUtf16

A lossy iterator over UTF-16 decoded chars of a string slice.

CharsLossyUtf32

A lossy iterator over UTF-32 decoded chars of a string slice.

CharsUtf16

An iterator over UTF-16 decoded chars of a string slice.

CharsUtf32

An iterator over UTF-32 decoded chars of a string slice.

Display

Helper struct for printing wide string values with format! and {}.

U16Str

16-bit wide string slice with undefined encoding.

U32Str

32-bit wide string slice with undefined encoding.

Type Definitions

WideStr

Alias for U16Str or U32Str depending on platform. Intended to match typical C wchar_t size on platform.