widestring

Module error

Errors returned by functions in this crate.

Structs

ContainsNul

An error returned to indicate that an invalid nul value was found in a string.

DecodeUtf16Error

An error that can be returned when decoding UTF-16 code points.

DecodeUtf32Error

An error that can be returned when decoding UTF-32 code points.

MissingNulTerminator

An error returned from to indicate that a terminating nul value was missing.

Utf16Error

Errors which can occur when attempting to interpret a sequence of u16 as UTF-16.

Utf32Error

Errors which can occur when attempting to interpret a sequence of u32 as UTF-32.

Enums

NulError

An error returned to indicate a problem with nul values occurred.