widestring

Module iter

Iterators for encoding and decoding slices of string data.

Structs

DecodeUtf16

An iterator that decodes UTF-16 encoded code points from an iterator of u16s.

DecodeUtf32

An iterator that decodes UTF-32 encoded code points from an iterator of u32s.

DecodeUtf16Lossy

An iterator that lossily decodes possibly ill-formed UTF-16 encoded code points from an iterator of u16s.

DecodeUtf32Lossy

An iterator that lossily decodes possibly ill-formed UTF-32 encoded code points from an iterator of u32s.

EncodeUtf8

An iterator that encodes an iterator of chars into UTF-8 bytes.

EncodeUtf16

An iterator that encodes an iterator of chars into UTF-16 u16 code units.

EncodeUtf32

An iterator that encodes an iterator of chars into UTF-32 u32 values.