widestring

Module ucstr

C-style wide string slices.

This module contains wide C string slices and related types.

Structs

CharIndicesLossyUtf16

A lossy iterator over the chars of a string slice, and their positions.

CharIndicesLossyUtf32

A lossy iterator over the chars of a string slice, and their positions.

CharIndicesUtf16

An iterator over the decoded chars of a string slice, and their positions.

CharIndicesUtf32

An iterator over the decoded chars of a string slice, and their positions.

CharsLossyUtf16

A lossy iterator over UTF-16 decoded chars of a string slice.

CharsLossyUtf32

A lossy iterator over UTF-32 decoded chars of a string slice.

CharsUtf16

An iterator over UTF-16 decoded chars of a string slice.

CharsUtf32

An iterator over UTF-32 decoded chars of a string slice.

Display

Helper struct for printing wide C string values with format! and {}.

U16CStr

C-style 16-bit wide string slice for U16CString.

U32CStr

C-style 32-bit wide string slice for U32CString.

Type Aliases

WideCStr

Alias for U16CStr or U32CStr depending on platform. Intended to match typical C wchar_t size on platform.