widestring

Module utfstr

UTF string slices.

This module contains UTF string slices and related types.

Structs

CharIndicesUtf16

An iterator over the chars of a string slice, and their positions

CharIndicesUtf32

An iterator over the chars of a string slice, and their positions

CharsUtf16

An iterator over the chars of a UTF-16 string slice

CharsUtf32

An iterator over the chars of a UTF-32 string slice

CodeUnits

An iterator over the u16 code units of a UTF-16 string slice

EscapeDebug

The return type of Utf16Str::escape_debug.

EscapeDefault

The return type of Utf16Str::escape_default.

EscapeUnicode

The return type of Utf16Str::escape_unicode.

Lines

An iterator over the lines of a crate::Utf16Str, crate::Utf32Str, or other wide string that has the char_indices method. Returns string slices.

Utf16Str

UTF-16 string slice for Utf16String.

Utf32Str

UTF-32 string slice for Utf32String.

Type Aliases

WideUtfStr

Alias for Utf16Str or Utf32Str depending on platform. Intended to match typical C wchar_t size on platform.